+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export * from './Box'
-export * from './ButtonV2'
+export * from './Button'
 export * from './Layout'
 export * from './SpringMotionBox'
 export * from './progressBar'
