+++ src/pages/Game/components/AlignedBox/index.tsx
@@ -1,9 +1,9 @@
-import { ButtonV2 } from '@/components'
+import { Button } from '@/components'
 import { GameMainAlign, GameMainCommon } from '@/constants/index'
 import { faSpaceShuttle } from '@fortawesome/free-solid-svg-icons'
 import { FC, HTMLAttributes, PropsWithChildren } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { PositionProps } from '../../Game.types'
+import type { PositionProps } from '../../Game.types'
 import * as S from './AlignedBox.styles'
 
 interface Props extends HTMLAttributes<HTMLDivElement>, PositionProps {}
@@ -17,9 +17,9 @@ export const AlignedArea: FC<PropsWithChildren<Props>> = ({ align, children, ...
       {children}
       <S.TitleAndButton align={align}>
         <h1>{GameMainAlign[align].titleText}</h1>
-        <ButtonV2 icon={faSpaceShuttle} bgColor="transparent" onClick={() => navigate(`${GameMainAlign[align].to}`)}>
+        <Button icon={faSpaceShuttle} bgColor="transparent" onClick={() => navigate(`${GameMainAlign[align].to}`)}>
           {GameMainCommon.buttonText}
-        </ButtonV2>
+        </Button>
       </S.TitleAndButton>
     </S.AlignedArea>
   )
