+++ src/pages/Game/components/SideOpenDescription/index.tsx
@@ -1,7 +1,7 @@
-import { Box } from '@/components/Box'
+import { Box } from '@/components'
 import { GameMainAlign, GameMainCommon } from '@/constants'
 import { FC, HTMLAttributes } from 'react'
-import { PositionProps } from '../../Game.types'
+import type { PositionProps } from '../../Game.types'
 import * as S from './SideOpenDescription.styles'
 
 interface Props extends HTMLAttributes<HTMLDivElement>, PositionProps {}
